# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 个人技术博客系统 API
  version: v1
  description: |
    个人技术博客系统 OpenAPI 规范。
    基础路径统一为 `/api/v1/**`，可直接导入 Apifox 使用。
servers:
  - url: http://localhost:8080
    description: 本地开发环境
...
paths:
...
  /api/v1/articles/{articleId}:
    get:
      tags:
        - Public Articles
      summary: 获取文章详情
      description: 权限级别 `PUBLIC`。仅可查看已发布文章详情。响应中的 `tags` 仅包含当前启用且仍存在的关联标签；历史关联的禁用标签或已被删除的标签不返回，不影响文章详情展示。文章所属二级分类及其父分类必须均存在且启用；分类不存在返回 `ARTICLE_CATEGORY_NOT_FOUND`，分类已禁用返回 `ARTICLE_CATEGORY_DISABLED`。文章关联作者必须存在；作者不存在返回 `ARTICLE_AUTHOR_NOT_FOUND`。P0 以 articleId 定位文章；slug 仅作为 P2 URL 可读化与 SEO 优化预留，不参与当前接口路径。
      operationId: getPublishedArticleDetail
      parameters:
        - $ref: '#/components/parameters/ArticleId'
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicArticleDetailResponse'
              example:
                code: 0
                message: 成功
                data:
                  id: 40001
                  title: Spring Boot 双 Token 登录实践
                  summary: 本文记录双 Token 的实现思路与接口设计
                  contentHtml: <h1>一、背景</h1><p>这里是文章正文</p>
                  coverUrl: https://cdn.example.com/cover/token.png
                  isTop: true
                  allowComment: true
                  viewCount: 128
                  commentCount: 0
                  likeCount: 0
                  favoriteCount: 0
                  publishedAt: '2026-04-22T23:00:00+08:00'
                  updatedAt: '2026-04-22T23:10:00+08:00'
                  category:
                    id: 21001
                    name: Java
                    level: 2
                    parent:
                      id: 20001
                      name: 技术
                  tags:
                    - id: 30001
                      name: JWT
                  author:
                    id: 10001
                    username: ccsanjuu
                    nickname: sanjuu
                    avatarUrl: https://cdn.example.com/avatar/1.png
                    bio: 专注后端和前端工程化
        '404':
          $ref: '#/components/responses/PublicArticleDetailNotFoundResponse'
        '409':
          $ref: '#/components/responses/PublicArticleDetailConflictResponse'
...
  /api/v1/admin/comments:
    get:
      tags:
        - Admin Comments
      summary: 获取后台评论分页列表
      description: 权限级别 `ADMIN`。支持按文章、作者、状态和评论层级筛选，默认按创建时间倒序。
      operationId: listAdminComments
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageNum'
        - $ref: '#/components/parameters/PageSize'
        - name: articleId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: userId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [PENDING, APPROVED, REJECTED, HIDDEN, DELETED]
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [TOP_LEVEL, REPLY]
        - name: createdAtFrom
          in: query
          description: 创建时间范围开始，ISO 8601 时间
          required: false
          schema:
            type: string
            format: date-time
        - name: createdAtTo
          in: query
          description: 创建时间范围结束，ISO 8601 时间
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCommentPageResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
  /api/v1/admin/comments/{commentId}/moderation:
    patch:
      tags:
        - Admin Comments
      summary: 审核、隐藏或删除评论
      description: 权限级别 `ADMIN`。管理员不能修改评论正文。REJECT、HIDE、DELETE 必须填写 reason；APPROVE 时清空历史处理原因。删除为逻辑删除，并级联逻辑删除全部后代。
      operationId: moderateComment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CommentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModerateCommentRequest'
      responses:
        '200':
          description: 处理成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentMutationResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '404':
          $ref: '#/components/responses/CommentNotFoundResponse'
        '409':
          $ref: '#/components/responses/CommentConflictResponse'
...
  /api/v1/admin/articles:
    get:
      tags:
        - Admin Articles
      summary: 获取后台文章分页列表
      description: 权限级别 `ADMIN`。支持按标题、分类、状态、置顶状态筛选。
      operationId: listAdminArticles
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageNum'
        - $ref: '#/components/parameters/PageSize'
        - name: title
          in: query
          description: 标题模糊搜索
          required: false
          schema:
            type: string
            example: Token
        - name: categoryId
          in: query
          description: 分类 ID。支持一级分类或二级分类筛选；文章实际绑定的仍是二级分类
          required: false
          schema:
            type: integer
            format: int64
            example: 20001
        - name: status
          in: query
          description: 文章状态
          required: false
          schema:
            type: string
            enum:
              - DRAFT
              - PUBLISHED
              - OFFLINE
            example: DRAFT
        - name: isTop
          in: query
          description: 是否置顶
          required: false
          schema:
            type: boolean
            example: true
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminArticlePageResponse'
              example:
                code: 0
                message: 成功
                data:
                  records:
                    - id: 40001
                      title: Spring Boot 双 Token 登录实践
                      summary: 本文记录双 Token 的实现思路与接口设计
                      status: DRAFT
                      isTop: true
                      coverUrl: https://cdn.example.com/cover/token.png
                      publishedAt: null
                      updatedAt: '2026-04-22T23:10:00+08:00'
                      category:
                        id: 21001
                        name: Java
                        level: 2
                        parent:
                          id: 20001
                          name: 技术
                  pageNum: 1
                  pageSize: 10
                  total: 1
                  totalPages: 1
                  hasNext: false
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
    post:
      tags:
        - Admin Articles
      summary: 创建文章
      description: 权限级别 `ADMIN`。创建新文章，仅允许直接保存为草稿或发布，不允许直接创建为下线状态。
      operationId: createAdminArticle
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleUpsertRequest'
            example:
              title: Spring Boot 双 Token 登录实践
              summary: 本文记录双 Token 的实现思路与接口设计
              contentMd: |
                # 一、背景

                这里是正文
              categoryId: 21001
              tagIds:
                - 30001
                - 30002
              coverUrl: https://cdn.example.com/cover/token.png
              isTop: true
              status: DRAFT
              allowComment: true
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleMutationResponse'
              example:
                code: 0
                message: 成功
                data:
                  id: 40001
                  status: DRAFT
                  publishedAt: null
                  createdAt: '2026-04-22T23:45:00+08:00'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '409':
          $ref: '#/components/responses/ArticleConflictResponse'
  /api/v1/admin/articles/{articleId}:
    get:
      tags:
        - Admin Articles
      summary: 获取后台文章详情
      description: 权限级别 `ADMIN`。用于后台编辑页回显完整文章内容。
      operationId: getAdminArticleDetail
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ArticleId'
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminArticleDetailResponse'
              example:
                code: 0
                message: 成功
                data:
                  id: 40001
                  title: Spring Boot 双 Token 登录实践
                  summary: 本文记录双 Token 的实现思路与接口设计
                  contentMd: |
                    # 一、背景

                    这里是正文
                  contentHtml: <h1>一、背景</h1><p>这里是正文</p>
                  contentText: 一、背景 这里是正文
                  coverUrl: https://cdn.example.com/cover/token.png
                  status: DRAFT
                  isTop: true
                  categoryId: 21001
                  tagIds:
                    - 30001
                    - 30002
                  allowComment: true
                  publishedAt: null
                  createdAt: '2026-04-22T22:50:00+08:00'
                  updatedAt: '2026-04-22T23:10:00+08:00'
        '404':
          $ref: '#/components/responses/ArticleNotFoundResponse'
    put:
      tags:
        - Admin Articles
      summary: 更新文章
      description: 权限级别 `ADMIN`。从未发布过的文章可保持草稿或发布；已发布过的文章仅允许在已发布与已下线之间流转，不允许回退为草稿。
      operationId: updateAdminArticle
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ArticleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleUpsertRequest'
            example:
              title: Spring Boot 双 Token 登录实践
              summary: 本文记录双 Token 的实现思路与接口设计
              contentMd: |
                # 一、背景

                这里是更新后的正文
              categoryId: 21001
              tagIds:
                - 30001
                - 30002
              coverUrl: https://cdn.example.com/cover/token.png
              isTop: true
              status: PUBLISHED
              allowComment: true
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleUpdateResponse'
              example:
                code: 0
                message: 成功
                data:
                  id: 40001
                  status: PUBLISHED
                  publishedAt: '2026-04-22T23:50:00+08:00'
                  updatedAt: '2026-04-22T23:50:00+08:00'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/ArticleNotFoundResponse'
        '409':
          $ref: '#/components/responses/ArticleConflictResponse'
    delete:
      tags:
        - Admin Articles
      summary: 删除文章
      description: 权限级别 `ADMIN`。
      operationId: deleteAdminArticle
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ArticleId'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoDataResponse'
              example:
                code: 0
                message: 成功
                data: null
        '404':
          $ref: '#/components/responses/ArticleNotFoundResponse'
  /api/v1/admin/articles/{articleId}/status:
    patch:
      tags:
        - Admin Articles
      summary: 修改文章状态
      description: 权限级别 `ADMIN`。从未发布过的文章仅允许发布；已发布过的文章仅允许在已发布与已下线之间流转。
      operationId: updateAdminArticleStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ArticleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArticleStatusRequest'
            example:
              status: OFFLINE
      responses:
        '200':
          description: 状态修改成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleStatusResponse'
              example:
                code: 0
                message: 成功
                data:
                  id: 40001
                  status: OFFLINE
                  updatedAt: '2026-04-22T23:55:00+08:00'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/ArticleNotFoundResponse'
        '409':
          $ref: '#/components/responses/ArticleConflictResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageNum:
      name: pageNum
      in: query
      description: 当前页码，从 1 开始，默认 1
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
    PageSize:
      name: pageSize
      in: query
      description: 每页条数，默认 10
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 20
        default: 10
        example: 10
...
    ArticleId:
      name: articleId
      in: path
      description: 文章 ID。P0 使用文章 ID 作为 URL 稳定定位标识，P2 可在前台 URL 中追加 slug 提升可读性。
      required: true
      schema:
        type: integer
        format: int64
        example: 40001
...
    CommentId:
      name: commentId
      in: path
      description: 评论 ID；回复接口中必须是顶层评论 ID
      required: true
      schema:
        type: integer
        format: int64
        example: 50001
...
  schemas:
    BaseResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: 业务状态码，`0` 表示成功
          example: 0
        message:
          type: string
          description: 业务描述信息
          example: 成功
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              nullable: true
              description: 错误场景通常为空
              example: null
    NoDataResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              nullable: true
              description: 无业务数据返回
              example: null
...
    PublicArticleDetail:
      type: object
      required:
        - id
        - title
        - summary
        - contentHtml
        - coverUrl
        - isTop
        - allowComment
        - viewCount
        - commentCount
        - likeCount
        - favoriteCount
        - publishedAt
        - updatedAt
        - category
        - tags
        - author
      properties:
        id:
          type: integer
          format: int64
          example: 40001
        title:
          type: string
          example: Spring Boot 双 Token 登录实践
        summary:
          type: string
          example: 本文记录双 Token 的实现思路与接口设计
        contentHtml:
          type: string
          description: 由 Markdown 转译得到的 HTML 正文，供前端渲染
          example: "<h1>一、背景</h1><p>这里是文章正文</p>"
        coverUrl:
          type: string
          example: https://cdn.example.com/cover/token.png
        isTop:
          type: boolean
          example: true
        allowComment:
          type: boolean
          example: true
        viewCount:
          type: integer
          example: 128
        commentCount:
          type: integer
          example: 0
        likeCount:
          type: integer
          example: 0
        favoriteCount:
          type: integer
          example: 0
        publishedAt:
          type: string
          format: date-time
          example: '2026-04-22T23:00:00+08:00'
        updatedAt:
          type: string
          format: date-time
          example: '2026-04-22T23:10:00+08:00'
        category:
          $ref: '#/components/schemas/SimpleCategory'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/SimpleTag'
        author:
          $ref: '#/components/schemas/SimpleAuthor'
    PublicArticleDetailResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PublicArticleDetail'
...
    AdminArticlePageData:
      allOf:
        - $ref: '#/components/schemas/PagedMeta'
        - type: object
          required:
            - records
          properties:
            records:
              type: array
              items:
                $ref: '#/components/schemas/AdminArticleListItem'
    AdminArticlePageResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AdminArticlePageData'
...
    AdminArticleDetailResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AdminArticleDetail'
    ArticleUpsertRequest:
      type: object
      required:
        - title
        - contentMd
        - categoryId
        - status
      properties:
        title:
          type: string
          example: Spring Boot 双 Token 登录实践
        summary:
          type: string
          example: 本文记录双 Token 的实现思路与接口设计
        contentMd:
          type: string
          description: Markdown 正文。`contentHtml` 与 `contentText` 由后端自动生成
          example: "# 一、背景\n\n这里是正文"
        categoryId:
          type: integer
          format: int64
          description: 二级分类 ID，文章不能直接绑定一级分类
          example: 21001
        tagIds:
          type: array
          items:
            type: integer
            format: int64
          example:
            - 30001
            - 30002
        coverUrl:
          type: string
          example: https://cdn.example.com/cover/token.png
        isTop:
          type: boolean
          example: true
        status:
          type: string
          description: 提交后的文章状态。创建时仅允许 `DRAFT` 或 `PUBLISHED`；更新时从未发布过的文章允许 `DRAFT` 或 `PUBLISHED`，已发布过的文章仅允许 `PUBLISHED` 或 `OFFLINE`。
          enum:
            - DRAFT
            - PUBLISHED
            - OFFLINE
          example: DRAFT
        allowComment:
          type: boolean
          example: true
    ArticleMutationData:
      type: object
      required:
        - id
        - status
        - publishedAt
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          example: 40001
        status:
          type: string
          enum:
            - DRAFT
            - PUBLISHED
            - OFFLINE
          example: DRAFT
        publishedAt:
          type: string
          format: date-time
          nullable: true
          example: null
        createdAt:
          type: string
          format: date-time
          example: '2026-04-22T23:45:00+08:00'
    ArticleMutationResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ArticleMutationData'
    ArticleUpdateData:
      type: object
      required:
        - id
        - status
        - publishedAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          example: 40001
        status:
          type: string
          enum:
            - DRAFT
            - PUBLISHED
            - OFFLINE
          example: PUBLISHED
        publishedAt:
          type: string
          format: date-time
          description: 首次发布时间。下线、重新发布或更新已发布文章时不刷新。
          nullable: true
          example: '2026-04-22T23:50:00+08:00'
        updatedAt:
          type: string
          format: date-time
          example: '2026-04-22T23:50:00+08:00'
    ArticleUpdateResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ArticleUpdateData'
    UpdateArticleStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          description: 目标状态。从未发布过的文章仅允许变为 `PUBLISHED`；已发布过的文章仅允许在 `PUBLISHED` 与 `OFFLINE` 间流转。
          enum:
            - DRAFT
            - PUBLISHED
            - OFFLINE
          example: OFFLINE
    ArticleStatusData:
      type: object
      required:
        - id
        - status
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          example: 40001
        status:
          type: string
          enum:
            - DRAFT
            - PUBLISHED
            - OFFLINE
          example: OFFLINE
        updatedAt:
          type: string
          format: date-time
          example: '2026-04-22T23:55:00+08:00'
    ArticleStatusResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ArticleStatusData'
...
    CommentStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, HIDDEN, DELETED]
...
    CommentMutationData:
      type: object
      required: [id, articleId, parentId, rootId, content, status, moderationReason, author, createdAt]
      properties:
        id:
          type: integer
          format: int64
        articleId:
          type: integer
          format: int64
        parentId:
          type: integer
          format: int64
          nullable: true
        rootId:
          type: integer
          format: int64
          nullable: true
        content:
          type: string
        status:
          $ref: '#/components/schemas/CommentStatus'
        moderationReason:
          type: string
          nullable: true
        author:
          $ref: '#/components/schemas/CommentAuthor'
        reviewedBy:
          type: integer
          format: int64
          nullable: true
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        deletedBy:
          type: integer
          format: int64
          nullable: true
        deletedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    CommentMutationResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/CommentMutationData'
...
    AdminCommentPageData:
      allOf:
        - $ref: '#/components/schemas/PagedMeta'
        - type: object
          required: [records]
          properties:
            records:
              type: array
              items:
                $ref: '#/components/schemas/AdminCommentItem'
    AdminCommentPageResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AdminCommentPageData'
    ModerateCommentRequest:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [APPROVE, REJECT, HIDE, DELETE]
        reason:
          type: string
          maxLength: 255
          description: REJECT、HIDE、DELETE 时必填；APPROVE 时不传并清空历史处理原因
...
  responses:
    BadRequestResponse:
      description: 请求参数不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 199001
            message: 请求参数不合法
            data: null
    UnauthorizedResponse:
      description: 未登录或 Access Token 无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 101001
            message: 未登录或 Access Token 无效
            data: null
    ForbiddenResponse:
      description: 无权限访问
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 101003
            message: 无权限访问
            data: null
...
    ArticleNotFoundResponse:
      description: 文章不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 103001
            message: 文章不存在
            data: null
    PublicArticleDetailNotFoundResponse:
      description: 文章不存在、不可见、关联分类不存在或关联作者不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            articleNotFound:
              summary: 文章不存在
              value:
                code: 103001
                message: 文章不存在
                data: null
            articleCategoryNotFound:
              summary: 文章分类不存在
              value:
                code: 103002
                message: 文章分类不存在
                data: null
            articleNotVisible:
              summary: 文章当前不可见
              value:
                code: 103008
                message: 文章当前不可见
                data: null
            articleAuthorNotFound:
              summary: 文章作者不存在
              value:
                code: 103009
                message: 文章作者不存在
                data: null
    PublicArticleDetailConflictResponse:
      description: 文章关联分类已禁用
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 103004
            message: 文章分类已禁用
            data: null
    ArticleConflictResponse:
      description: 文章关联资源冲突或状态流转不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 103007
            message: 文章状态流转不合法
            data: null
...
    CommentNotFoundResponse:
      description: 评论不存在、已删除或当前用户不可见
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 106001
            message: 评论不存在
            data: null
...
    CommentConflictResponse:
      description: 评论业务状态冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            commentDisabled:
              value:
                code: 106002
                message: 文章已关闭评论
                data: null
            parentUnavailable:
              value:
                code: 106003
                message: 回复目标不可用
                data: null
            statusConflict:
              value:
                code: 106006
                message: 评论状态流转不合法
                data: null
...